changed to float

+++ main.js
@@ -41,7 +41,7 @@ const lat = columns[7];
 const lon = columns[8];
 lats.push(lat);
 lons.push(lon);
-L.marker([parseInt(lat),parseInt(lon)]).addTo(map);
+L.marker([parseFloat(lat),parseFloat(lon)]).addTo(map);
 });
 
 console.log(lats[10]);
